+++ src/board/board/board.js
@@ -1,0 +1,9 @@
+// more of a placeholder for now - 
+// should talk when merged to properly integrate this with 
+// existing "board" and "players" implementations if any exist
+function Board(players) { 
+  this.players = players; 
+  this.spaces = [];
+} 
+
+module.exports = Board;
