+++ src/board/jail/goToJail.js
@@ -1,0 +1,5 @@
+
+function GoToJail() {}
+GoToJail.prototype = Object.create(Space.prototype);
+GoToJail.prototype.performLandingAction = function(player) { 
+};
