+++ src/board/other/chance.js
@@ -1,0 +1,6 @@
+function Chance() {} 
+Chance.prototype = Object.create(Space.prototype);
+Chance.prototype.performLandingAction = function(player) { 
+  // todo 
+  // call andrew's function relating to chance cards
+};
