+++ src/board/other/communityChest.js
@@ -1,0 +1,6 @@
+function CommunityChest() {}
+CommunityChest.prototype = Object.create(Space.prototype);
+CommunityChest.prototype.performLandingAction = function(player) { 
+  // todo 
+  // see above chance() 
+};
