+++ src/board/other/freeParking.js
@@ -1,0 +1,5 @@
+function FreeParking() {}
+FreeParking.prototype = Object.create(Space.prototype);
+FreeParking.prototype.performLandingAction = function(player) { 
+  // nothing! 
+};
