+++ src/board/other/go.js
@@ -1,0 +1,3 @@
+function Go() {}
+Go.prototype = Object.create(Space.prototype);
+
