+++ src/board/properties.js
@@ -40,139 +40,6 @@ var PG_BLUE = 7;
 var PG_RR = 8;
 var PG_UTIL = 9;
 
-// placeholder for now - should talk when merged to properly integrate this with 
-// existing "board" and "players" implementations
-function Board(players) { 
-  this.players = players; // just a placeholder for now - on merge should combine 
-  this.spaces = [];
-} 
-
-// class definitions for any space on the board - 
-// properties, utilities, free parking, go, etc. 
-function Space(name) { 
-  this.name = name; 
-  this.occupier = null;
-}
-// every space needs a landing action
-Space.prototype.performLandingAction = function(){
-  console.log("You landed on " + this.name);
-}; 
-
-function Property(name, cost, propertyGroup) { 
-  Space.call(this, name);
-  this.cost = cost;
-  this.mortage = .5*cost;
-  this.propertyGroup = propertyGroup; // see PG_X constants 
-  this.owner = null;
-}
-Property.prototype = Object.create(Space.prototype); // subclassing space
-Property.prototype.performLandingAction = function(player) { 
-
-  if (self.owner == player) return; 
-
-  // todo  - finish hashing this out 
-
-}; 
-
-// rent should be array with following format: 
-// [1 owned (rent), 2 owned, 3 owned, 4 owned]
-function RailroadProperty(name, cost, propertyGroup, rent) { 
-  this.rent = rent;
-  Property.call(this, name, cost, propertyGroup);
-}
-RailroadProperty.prototype = Object.create(Property.prototype);
-RailroadProperty.prototype.performLandingAction = function(player) {
-  // todo
-  // probably call something like: 
-  // handlePropertyLanding(name, cost, rent, player, multiplier);
-}; 
-
-// rent is assumed here (4x, 10x) 
-function UtilityProperty(name, cost, propertyGroup) { 
-  Property.call(this, name, cost, propertyGroup);
-}
-UtilityProperty.prototype = Object.create(Property.prototype);
-UtilityProperty.prototype.performLandingAction = function(player, dice) { 
-  // todo - need dice access here
-}; 
-
-// rent should be array with following format: 
-// [rent, 1 house, 2 houses, 3 houses, 4 houses, hotel] 
-function HousingProperty(name, cost, propertyGroup, rent) { 
-  this.rent = rent;
-  this.numHouses = 0;
-  Property.call(this, name, cost, propertyGroup);
-}
-HousingProperty.prototype = Object.create(Property.prototype);
-HousingProperty.prototype.performLandingAction = function(player) { 
-  // todo
-  // see above railroad property fn comments
-};
-
-
-function Chance() {} 
-Chance.prototype = Object.create(Space.prototype);
-Chance.prototype.performLandingAction = function(player) { 
-  // todo 
-  // call andrew's function relating to chance cards
-}; 
-
-function CommunityChest() {}
-CommunityChest.prototype = Object.create(Space.prototype);
-CommunityChest.prototype.performLandingAction = function(player) { 
-  // todo 
-  // see above chance() 
-}; 
-
-function FreeParking() {}
-FreeParking.prototype = Object.create(Space.prototype);
-FreeParking.prototype.performLandingAction = function(player) { 
-  // nothing! 
-}; 
-
-function Jail() {}
-GoToJail.prototype = Object.create(Space.prototype);
-GoToJail.prototype.performLandingAction = function(player) { 
-  //  player.go_to_jail(); todo: make this fn in player
-}; 
-
-function GoToJail() {}
-GoToJail.prototype = Object.create(Space.prototype);
-GoToJail.prototype.performLandingAction = function(player) { 
-}; 
-
-function Tax() { 
-  this.taxAmount = 0; // give default value of 0, should be overridden 
-}
-Tax.prototype = Object.create(Space.prototype);
-Tax.prototype.performLandingAction = function(player) { 
-  // should probably refactor this into a fn
-  // that checks balance, prompts player to mortage/sell
-  // before declaring him bankrupt
-  this.player.money -= this.taxAmount;
-}; 
-
-function IncomeTax() {
-  this.taxAmount = 200; // flat tax - will give user option later
-}
-IncomeTax.prototype = Object.create(Tax.prototype);
-IncomeTax.prototype.performLandingAction = function(player) { 
-  // todo: prompt user for choice here
-  // if (userWantsFlatTax) this.taxAmount = 200; 
-  // else this.taxAmount = playerAssetValue(player)*.10;
-}; 
-
-function LuxuryTax() { 
-  this.taxAmount = 75;
-}
-LuxuryTax.prototype = Object.create(Tax.prototype);
-IncomeTax.prototype.performLandingAction = function(player) { 
-  // todo
-}; 
-
-function Go() {}
-Go.prototype = Object.create(Space.prototype);
-
 // end of space class definitions
