+++ src/board/taxes/luxuryTax.js
@@ -1,0 +1,7 @@
+function LuxuryTax() { 
+  this.taxAmount = 75;
+}
+LuxuryTax.prototype = Object.create(Tax.prototype);
+IncomeTax.prototype.performLandingAction = function(player) { 
+  // todo
+};
